+++ frontend/src/lib/components/settings/LeaveVisibleInfo.svelte
@@ -1,0 +1,5 @@
+<p>
+    Normally after leaving a game TFView will display a "No game is currently active" popup
+    which hides the players that were in your last game. This setting removes that popup,
+    allowing you to easily interact with players from a game you left.
+</p>