+++ .github/release-notes.md
@@ -2,3 +2,4 @@
 * Improved syntax for calling scripter arguments to be more reliable
     * While the old tfivew.script(arg1,arg2) syntax still works, tfview.script[arg1|arg2] is now recommended
 * Added icon for players that you had muted when TFView was launched
+* Added setting to leave players from the last game you were in visible before you load into the next one